⚡️ perf(game): 优化游戏循环和帧率计算
- 【性能优化】将fps和frameTime数据类型从int改为float，提高计算精度
- 【性能优化】改进帧率计算逻辑，使用浮点数运算避免整数除法精度损失
- 【性能优化】添加帧率统计功能，每500ms计算一次平均帧率
- 【性能优化】优化游戏循环中的时间计算，使用accumulator模式提高稳定性

+++ src/Game.h
@@ -64,8 +64,8 @@ private:
     int windowHeight = 960;
     bool isRunning = false;
 
-    int fps = 60;
-    Uint32 frameTime = 0;
+    float fps = 60.f;
+    float frameTime = 0;
     float deltaTime = 0;
 };
